Move cache_callback to dependencies
